The great splitting: Break apart classes into their own files
