+++ src/main/scala/I18N.scala
@@ -1,0 +1,12 @@
+package org.levelspace
+
+import java.util.{ ResourceBundle, MissingResourceException }
+
+object I18N {
+  private val bundle = ResourceBundle.getBundle("XW_LS_Strings")
+  def get(key: String) = try {
+    bundle.getString(key)
+  } catch {
+    case e: MissingResourceException => key
+  }
+}
